gbfs

+++ src/pathfinder/Pathfinder.java
@@ -1,10 +1,9 @@
 package pathfinder;
 
+import element.Board;
 import java.util.HashMap;
 import java.util.List;
 
-import element.Board;
-
 public interface Pathfinder {
-    public List<Board> search(Board board, HashMap<Integer, Boolean> visited);
+    public List<Board> search(Board board, HashMap<Long, Boolean> visited);
 }
